Finally deprecating UpdateIDs.py, and the IDs page.
Added a better-functioning function to get the date from days-in, after realizing that all the IDs page really had was the ID from the NYT page, which no one uses and I was only printing to the user because I could. The date and the number of the game is much more important, and I no longer need to read from another file, etc. Much better.